% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbrAPI.R
\docType{package}
\name{cbrAPI}
\alias{cbrAPI}
\alias{cbrAPI-package}
\title{Allows to download and read dbf-files with different forms of accounting reports (101, 102, 123, 134, 135) for a certain time period, as well as find information about the banks on Central Bank of Russia site.}
\description{
The goal of this project is to make available to users all over the world the unique data of the accounting statements of Russian banks, which is collected by the Central Bank of Russian Federation. Data is represented as archived dbf files (database files). The package provides the opportunity to download these archives from the Central Bank's website, read them and load into R. You can download any of the five forms of the accounting reports (101, 102, 123, 134, 135) for any given time period or over the entire available. You can also get detailed information on any Bank, knowing his registration number to find out the name, address, is it registered it in the Deposit Insurance Agency, is it went bankrupt or not.
}

