% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbrAPI.R
\name{getBankInfo}
\alias{getBankInfo}
\title{Get information about Russian banks}
\usage{
getBankInfo(regnum)
}
\arguments{
\item{regnum}{Registration number of some bank. Can be integer or string.}
}
\value{
Returns data frame with information about bank.\cr\cr
\code{name} - Bank's name\cr
\code{regnum} - Bank's registration number\cr
\code{address} - Bank's address\cr
\code{balance} - Presence of balance reports (monthly)\cr
\code{profitandlosses} - Presence of profit and loss statement (quarterly)\cr
\code{dia} - Is this bank registered in the Deposit Insurance Agency (DIA)?\cr
\code{intheprocess} - Is this bank in the process of registration (no license yet)?\cr
\code{withdrawed} - Is licence withdrawn?\cr
\code{revoked} - Is licence revoked?\cr
\code{eliminated} - Is licence eliminated?
}
\description{
This function allows you to parse information about Russian banks by registration number from site of Central Bank of Russia (www.cbr.ru). \cr
See \code{vignette(package = 'cbrAPI')} for proper use.
}
\examples{
info <- getBankInfo(2856)
info <- getBankInfo("30")$name
}
\seealso{
\code{\link{readHTMLTable}} for parsing html tables, \code{\link{download.dbf}} for downloading financial statements of Russian banks
}
\keyword{accounting}
\keyword{banks}
\keyword{cbr}
\keyword{dbf}
\keyword{download}
\keyword{read}
\keyword{reports}
\keyword{russia}

