% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cbrAPI.R
\name{download.dbf.all}
\alias{download.dbf.all}
\title{Download financial statements of Russian banks}
\usage{
download.dbf.all(form = 101, tofactor = FALSE, savefiles = FALSE)
}
\arguments{
\item{form}{Report form, contains what type of form to download. Can be \code{101}, \code{102}, \code{123}, \code{134} and \code{135}. Defaults to \code{101}.}

\item{tofactor}{Convert resulted data frame to factor or not? Defaults to \code{FALSE}.}

\item{savefiles}{Save downloaded dbf-files or not? Defaults to \code{FALSE}.}
}
\value{
Returns data frame with row-binded content of dbf-files
}
\description{
This function allows you to download and read dbf-files FOR ALL AVAILABLE DATES, which contain all available accounting reports and financial statements from site of Central Bank of Russia (www.cbr.ru) \cr
See \code{vignette(package = 'cbrAPI')} for proper use.
}
\examples{
data135 <- download.dbf.all(form=135)
}
\seealso{
\code{\link{download.file}} for simple downloading, \code{\link{read.dbf}} for reading dbf-files
}
\keyword{accounting}
\keyword{banks}
\keyword{cbr}
\keyword{dbf}
\keyword{download}
\keyword{read}
\keyword{reports}
\keyword{russia}

